Methods

Index

PhysicalParticles.ZeroValue

PhysicalParticles.assign_particles

PhysicalParticles.astro

PhysicalParticles.average

PhysicalParticles.averagebymass

PhysicalParticles.axisunit

PhysicalParticles.center

PhysicalParticles.center_x

PhysicalParticles.center_y

PhysicalParticles.center_z

PhysicalParticles.cgs

PhysicalParticles.cylinderial

PhysicalParticles.cylinderial2xyz

PhysicalParticles.datadimension

PhysicalParticles.distance

PhysicalParticles.extent

PhysicalParticles.getuAcc

PhysicalParticles.getuAmount

PhysicalParticles.getuCurrent

PhysicalParticles.getuDensity

PhysicalParticles.getuDensity2D

PhysicalParticles.getuEnergy

PhysicalParticles.getuEntropy

PhysicalParticles.getuLength

PhysicalParticles.getuLuminosity

PhysicalParticles.getuMass

PhysicalParticles.getuPressure

PhysicalParticles.getuTemperature

PhysicalParticles.getuTime

PhysicalParticles.getuVel

PhysicalParticles.getunits

PhysicalParticles.mass_center

PhysicalParticles.max_x

PhysicalParticles.max_y

PhysicalParticles.max_z

PhysicalParticles.maximum_x

PhysicalParticles.maximum_y

PhysicalParticles.maximum_z

PhysicalParticles.min_x

PhysicalParticles.min_y

PhysicalParticles.min_z

PhysicalParticles.minimum_x

PhysicalParticles.minimum_y

PhysicalParticles.minimum_z

PhysicalParticles.pconvert

PhysicalParticles.pos_center

PhysicalParticles.rand_pvector

PhysicalParticles.rand_pvector2d

PhysicalParticles.randn_pvector

PhysicalParticles.randn_pvector2d

PhysicalParticles.rotate

PhysicalParticles.rotate_x

PhysicalParticles.rotate_y

PhysicalParticles.rotate_z

PhysicalParticles.si

PhysicalParticles.spherical

PhysicalParticles.spherical2xyz

PhysicalParticles.split_data

Public

Missing docstring.

Missing docstring for length. Check Documenter's build log for details.

PhysicalParticles.split_data — Function

split_data(data::Array, i::Int64, N::Int64)

split data to N sections, return the ith section

Examples

julia> split_data([1,2,3], 1, 2)
2-element Vector{Int64}:
 1
 2

julia> split_data([1,2,3], 2, 2)
1-element Vector{Int64}:
 3

julia> split_data([1,2,3], 3, 4)
1-element Vector{Int64}:
 3

PhysicalParticles.extent — Function

extent(a::Array{T, N})
extent(data::Dict)
extent(a::Extent2D, b::Extent2D)
extent(a::Extent, b::Extent)

Get Extent of an array of mathical vectors or particles, or by comparing two Extents and even array of Extents

Examples

julia> extent([Ball(PVector(-1.0u"m", 1.0u"m", 1.0u"m"), PVector(u"m/s"), PVector(u"m/s^2"), 1.0u"kg", 1),
               Ball(PVector(1.0u"m", -1.0u"m", -1.0u"m"), PVector(u"m/s"), PVector(u"m/s^2"), 1000.0u"g", 2)])
Extent: xMin = -1.0 m, xMax = 1.0 m, yMin = -1.0 m, yMax = 1.0 m, zMin = -1.0 m, zMax = 1.0 m, SideLength = 2.0 m, Center = PVector(0.0 m, 0.0 m, 0.0 m)

Missing docstring.

Missing docstring for volume. Check Documenter's build log for details.

Missing docstring.

Missing docstring for area. Check Documenter's build log for details.

PhysicalParticles.distance — Function

function distance(a::AbstractPoint, b::AbstractPoint)
distance(a::AbstractParticle, b::AbstractParticle)

Euclidean distance between two points or two particles.

PhysicalParticles.average — Function

function average(data, symbol::Symbol)

Average on field symbol of elements in an array or dict of arrays

Examples

a = rand_pvector(5)
average(a, :x)

PhysicalParticles.averagebymass — Function

function averagebymass(data, symbol::Symbol)
function averagebymass(data::StructArray, symbol::Symbol)

Average on field symbol weighted by :Mass of elements in an array of dict of arrays

Examples

averagebymass(data, :Pos)

PhysicalParticles.pos_center — Function

pos_center(a::Array{T}) where T <: AbstractParticle
pos_center(a::StructArray)

Compute averaged position with equal weights. It is different from middle(a, :Pos) which computes the middle value of a symbol (useful to avoid influences from distant particles).

There are differences among center, pos_center, mass_center and median:

center: box center of particles

pos_center: average position of particles

mass_center: mass weighted average position of particles

median: middle value of positions of particles

PhysicalParticles.mass_center — Function

mass_center(a::Array{T}) where T <: AbstractParticle
mass_center(a::StructArray)

Compute mass center of particles, which is weighted by mass.

There are differences among center, pos_center, mass_center and median:

center: box center of particles

pos_center: average position of particles

mass_center: mass weighted average position of particles

median: middle value of positions of particles

PhysicalParticles.pconvert — Function

pconvert(a::Abstractay{T,1}) where T<:Number

convert two-element array to PVector2D, and three-element array to PVector

Examples

julia> pconvert([1.0, 2.0])
PVector2D{Float64}(1.0, 2.0)

julia> pconvert([1.0, 2.0, 3.0])
PVector{Float64}(1.0, 2.0, 3.0)

pconvert(a::Abstractay{T,2}) where T<:Number

convert 2xN Array to Array{PVector2D,1}, 3xN Array to Array{PVector,1}

Examples

julia> pconvert([1.0 3.0; 
                 2.0 4.0])
2-element Vector{PVector2D}:
 PVector2D{Float64}(1.0, 2.0)
 PVector2D{Float64}(3.0, 4.0)

julia> pconvert([1.0 4.0;
                 2.0 5.0;
                 3.0 6.0])
2-element Vector{PVector}:
 PVector{Float64}(1.0, 2.0, 3.0)
 PVector{Float64}(4.0, 5.0, 6.0)

PhysicalParticles.astro — Function

Set default units to uAstro

PhysicalParticles.cgs — Function

Set default units to uCGS

PhysicalParticles.si — Function

Set default units to uSI

PhysicalParticles.getunits — Function

getunits(units = uDefaults)

Return Tuple of units from units

PhysicalParticles.getuLength — Function

function getuLength(::Nothing)
function getuLength(units)

Extract length unit from units or nothing

Examples

julia> getuLength(nothing)

julia> getuLength(uAstro)
kpc

julia> getuLength(uSI)
m

julia> getuLength(uCGS)
cm

PhysicalParticles.getuTime — Function

function getuTime(::Nothing)
function getuTime(units)

Extract time unit from units or nothing

Examples

julia> getuTime(nothing)

julia> getuTime(uAstro)
Gyr

julia> getuTime(uSI)
s

julia> getuTime(uCGS)
s

PhysicalParticles.getuCurrent — Function

function getuCurrent(::Nothing)
function getuCurrent(units)

Extract electric current unit from units or nothing

Examples

julia> getuCurrent(nothing)

julia> getuCurrent(uAstro)
A

julia> getuCurrent(uSI)
A

julia> getuCurrent(uCGS)
A

PhysicalParticles.getuTemperature — Function

function getuTemperature(::Nothing)
function getuTemperature(units)

Extract temperature unit from units or nothing

Examples

julia> getuTemperature(nothing)

julia> getuTemperature(uAstro)
K

julia> getuTemperature(uSI)
K

julia> getuTemperature(uCGS)
K

PhysicalParticles.getuLuminosity — Function

function getuLuminosity(::Nothing)
function getuLuminosity(units)

Extract luminosity unit from units or nothing

Examples

julia> getuLuminosity(nothing)

julia> getuLuminosity(uAstro)
cd

julia> getuLuminosity(uSI)
cd

julia> getuLuminosity(uCGS)
cd

PhysicalParticles.getuMass — Function

function getuMass(::Nothing)
function getuMass(units)

Extract mass unit from units or nothing

Examples

julia> getuMass(nothing)

julia> getuMass(uAstro)
M⊙

julia> getuMass(uSI)
kg

julia> getuMass(uCGS)
g

PhysicalParticles.getuAmount — Function

function getuAmount(::Nothing)
function getuAmount(units)

Extract amount unit from units or nothing

Examples

julia> getuAmount(nothing)

julia> getuAmount(uAstro)
mol

julia> getuAmount(uSI)
mol

julia> getuAmount(uCGS)
mol

PhysicalParticles.getuVel — Function

function getuVel(::Nothing)
function getuVel(units)

Extract velocity unit from units or nothing

Examples

julia> getuVel(nothing)

julia> getuVel(uAstro)
kpc Gyr^-1

julia> getuVel(uSI)
m s^-1

julia> getuVel(uCGS)
cm s^-1

PhysicalParticles.getuAcc — Function

function getuAcc(::Nothing)
function getuAcc(units)

Extract acceleration unit from units or nothing

Examples

julia> getuAcc(nothing)

julia> getuAcc(uAstro)
kpc Gyr^-2

julia> getuAcc(uSI)
m s^-2

julia> getuAcc(uCGS)
cm s^-2

PhysicalParticles.getuEnergy — Function

function getuEnergy(::Nothing)
function getuEnergy(units)

Extract energy unit from units or nothing

Examples

julia> getuEnergy(nothing)

julia> getuEnergy(uAstro)
kpc^2 M⊙ Gyr^-2

julia> getuEnergy(uSI)
kg m^2 s^-2

julia> getuEnergy(uCGS)
g cm^2 s^-2

PhysicalParticles.getuEntropy — Function

function getuEntropy(::Nothing)
function getuEntropy(units)

Extract entropy unit from units or nothing

Examples

julia> getuEntropy(nothing)

julia> getuEntropy(uAstro)
kpc^2 M⊙ K^-1 Gyr^-2

julia> getuEntropy(uSI)
kg m^2 K^-1 s^-2

julia> getuEntropy(uCGS)
g cm^2 K^-1 s^-2

PhysicalParticles.getuDensity — Function

function getuDensity(::Nothing)
function getuDensity(units)

Extract 3D density unit from units or nothing

Examples

julia> getuDensity(nothing)

julia> getuDensity(uAstro)
M⊙ kpc^-3

julia> getuDensity(uSI)
kg m^-3

julia> getuDensity(uCGS)
g cm^-3

PhysicalParticles.getuDensity2D — Function

function getuDensity2D(::Nothing)
function getuDensity2D(units)

Extract 2D density unit from units or nothing

Examples

julia> getuDensity2D(nothing)

julia> getuDensity2D(uAstro)
M⊙ kpc^-2

julia> getuDensity2D(uSI)
kg m^-2

julia> getuDensity2D(uCGS)
g cm^-2

PhysicalParticles.getuPressure — Function

function getuPressure(::Nothing)
function getuPressure(units)

Extract 2D density unit from units or nothing

Examples

julia> getuPressure(nothing)

julia> getuPressure(uAstro)
M⊙ kpc^-1 Gyr^-2

julia> getuPressure(uSI)
kg m^-1 s^-2

julia> getuPressure(uCGS)
g cm^-1 s^-2

PhysicalParticles.axisunit — Function

axisunit(::Nothing)
axisunit(u::Units)
axisunit(s::AbstractString, u::Units)

Return a String for pretty printing.

Examples

julia> axisunit(nothing)
""

julia> axisunit(u"m")
" [m]"

julia> axisunit("Time", u"Gyr")
"Time [Gyr]"

PhysicalParticles.ZeroValue — Type

ZeroValue(::Type{T}, ::Nothing) where T<:Number
ZeroValue(::Type{T}, units::Vector{Unitful.FreeUnits{N, D, nothing} where D where N} = uAstro) where T<:Number

Construct an immutable struct providing zero values of type T in corresponding units (default is uAstro). Use nothing if unitless. Useful for accumulated summation, array initialization, etc.

Examples

ZeroValue(Float32, nothing)
ZeroValue(Int32)
ZeroValue(BigFloat, uSI)
ZeroValue(Measurement, uCGS)

function ZeroValue(::Nothing)
function ZeroValue(units::Vector{Unitful.FreeUnits{N, D, nothing} where D where N} = uAstro)

Construct an immutable struct providing zero Measurement values in corresponding units (default is uAstro). Use nothing if unitless. Useful for accumulated summation, array initialization, etc.

Examples

ZeroValue(nothing)
ZeroValue()
ZeroValue(uSI)
ZeroValue(uCGS)

PhysicalParticles.rand_pvector — Function

rand_pvector2d(n::Integer, T::DataType = Float64)
rand_pvector2d(n::Integer, u::Units, T::DataType = Float64)
rand_pvector(n::Integer, T::DataType = Float64)
rand_pvector(n::Integer, u::Units, T::DataType = Float64)

Generate uniformly distributed PVectors

Examples

julia> p = rand_pvector(5)
julia> pu = rand_pvector2d(5, u"m")

PhysicalParticles.rand_pvector2d — Function

rand_pvector2d(n::Integer, T::DataType = Float64)
rand_pvector2d(n::Integer, u::Units, T::DataType = Float64)
rand_pvector(n::Integer, T::DataType = Float64)
rand_pvector(n::Integer, u::Units, T::DataType = Float64)

Generate uniformly distributed PVectors

Examples

julia> p = rand_pvector(5)
julia> pu = rand_pvector2d(5, u"m")

PhysicalParticles.randn_pvector — Function

randn_pvector2d(n::Integer, T::DataType = Float64)
randn_pvector2d(n::Integer, u::Units, T::DataType = Float64)
randn_pvector(n::Integer, T::DataType = Float64)
randn_pvector(n::Integer, u::Units, T::DataType = Float64)

Generate uniformly distributed PVectors

Examples

julia> p = randn_pvector(5)
julia> pu = randn_pvector2d(5, u"m")

PhysicalParticles.randn_pvector2d — Function

randn_pvector2d(n::Integer, T::DataType = Float64)
randn_pvector2d(n::Integer, u::Units, T::DataType = Float64)
randn_pvector(n::Integer, T::DataType = Float64)
randn_pvector(n::Integer, u::Units, T::DataType = Float64)

Generate uniformly distributed PVectors

Examples

julia> p = randn_pvector(5)
julia> pu = randn_pvector2d(5, u"m")

PhysicalParticles.assign_particles — Function

assign_particles(particles::Array{P,N} where P<:AbstractParticle, symbol::Symbol, data::Array) where N

Assign the symbol of particles with elements in data one by one. particles and data must have equal length.

Examples

julia> assign_particles([Ball(uSI) for i=1:3], :Pos, rand_pvector(3, u"m"))

assign_particles(particles::Array{P,N} where P<:AbstractParticle, symbol::Symbol, data) where N
assign_particles(particles::StructArray, symbol::Symbol, data)

Assign the symbol of particles to data identically.

Examples

julia> assign_particles([Ball(uSI) for i=1:3], :Mass, 1.0u"kg")

PhysicalParticles.datadimension — Function

function datadimension

Traits function to specify whether the data is unitful/unitless and 2D/3D.

Examples

julia> datadimension(PVector2D())
Unitless2D()

julia> datadimension(PVector())
Unitless3D()

julia> datadimension(Star2D(uAstro))
Physical2D()

julia> datadimension(Star(uAstro))
Physical3D()

julia> datadimension([PVector2D()])
Unitless2D()

julia> datadimension([PVector()])
Unitless3D()

julia> datadimension([Star2D(uAstro)])
Physical2D()

julia> datadimension([Star(uAstro)])
Physical3D()

PhysicalParticles.center — Function

Compute box center of points or particles

There are differences among center, pos_center, mass_center and median:

center: box center of particles

pos_center: average position of particles

mass_center: mass weighted average position of particles

median: middle value of positions of particles

PhysicalParticles.center_x — Function

Compute box center of x direction of points or particles

PhysicalParticles.center_y — Function

center_y(a)

Compute box center of y direction of points or particles

PhysicalParticles.center_z — Function

Compute box center of z direction of points or particles

PhysicalParticles.rotate_x — Function

rotate_x(p::PVector, roll::Number)

Rotate p around x-axis by roll in radian angle. The rotation direction is right-handed, which is counter-clockwise in y-z plane. Angles can have units, for example, u"°" (\degree).

The rotation matrix:

           | 1      0          0      |
Rx(roll) = | 0  cos(roll)  -sin(roll) |
           | 0  sin(roll)   cos(roll) |

rotate_x(data, roll::Number; vel::Bool = true)

Rotate all particles in data around x-axis by roll in radian angle.

vel::Bool = false: Rotate velocity around origin at the same time.

rotate_x(p::PVector, roll::Number, center::PVector)

Rotate p around the x-axis of center point by roll in radian angle.

rotate_x(data, roll::Number, center::PVector; vel::Bool = true)

Rotate all particles in data around the x-axis of center point by roll in radian angle.

vel::Bool = false: Rotate velocity around origin at the same time.

PhysicalParticles.rotate_y — Function

rotate_y(p::PVector, pitch::Number)

Rotate p around y-axis by pitch in radian angle. The rotation direction is right-handed, which is counter-clockwise in y-z plane. Angles can have units, for example, u"°" (\degree).

The rotation matrix:

            |  cos(pitch)   0   sin(pitch) |
Ry(pitch) = |       0       1        0     |
            | -sin(pitch)   0   cos(pitch) |

rotate_y(data, pitch::Number; vel::Bool = true)

Rotate all particles in data around y-axis by pitch in radian angle.

vel::Bool = false: Rotate velocity around origin at the same time.

rotate_y(p::PVector, pitch::Number, center::PVector)

Rotate p around the y-axis of center point by pitch in radian angle.

rotate_y(data, pitch::Number, center::PVector; vel::Bool = true)

Rotate all particles in data around the y-axis of center point by pitch in radian angle.

vel::Bool = false: Rotate velocity around origin at the same time.

PhysicalParticles.rotate_z — Function

rotate_z(p::PVector, yaw::Number)

Rotate p around z-axis by yaw in radian angle. The rotation direction is right-handed, which is counter-clockwise in y-z plane. Angles can have units, for example, u"°" (\degree).

The rotation matrix:

          | cos(yaw)  -sin(yaw)  0 |
Rz(yaw) = | sin(yaw)   cos(yaw)  0 |
          |    0          0      1 |

rotate_z(data, yaw::Number; vel::Bool = true)

Rotate all particles in data around z-axis by yaw in radian angle.

vel::Bool = false: Rotate velocity around origin at the same time.

rotate_z(p::PVector, yaw::Number, center::PVector)

Rotate p around the z-axis of center point by yaw in radian angle.

rotate_z(data, yaw::Number, center::PVector; vel::Bool = true)

Rotate all particles in data around the z-axis of center point by yaw in radian angle.

vel::Bool = false: Rotate velocity around origin at the same time.

PhysicalParticles.rotate — Function

rotate(p::PVector, α::Number, β::Number, γ::Number)

Rotate p by radian Euler angles (α, β, γ) = (roll, pitch, yaw). Angles can have units, for example, u"°" (\degree).

rotate(data, yaw::Number, center::PVector; vel::Bool = true)

Rotate all particles in data by radian Euler angles (α, β, γ) = (roll, pitch, yaw).

vel::Bool = false: Rotate velocity around origin at the same time.

rotate(p::PVector, α::Number, β::Number, γ::Number, center::PVector)

Rotate p by radian Euler angles (α, β, γ) = (roll, pitch, yaw) around center point. Angles can have units, for example, u"°" (\degree).

rotate(data, yaw::Number, center::PVector; vel::Bool = true)

Rotate all particles in data by radian Euler angles (α, β, γ) = (roll, pitch, yaw) around center point.

vel::Bool = false: Rotate velocity around origin at the same time.

rotate(p::PVector, vec::PVector, θ::Number)

Rotate p around direction vector vec by angle θ.

For normalized vec = (x, y, z), the rotation matrix:

            |  cos(θ) + (1-cos(θ))*x^2   (1-cos(θ))*x*y - sin(θ)*z  (1-cos(θ))*x*z + sin(θ)*y |
M(vec, θ) = | (1-cos(θ))*y*x + sin(θ)*z   cos(θ) + (1-cos(θ))*y^2   (1-cos(θ))*y*z - sin(θ)*x |
            | (1-cos(θ))*z*x - sin(θ)*y  (1-cos(θ))*z*y + sin(θ)*x   cos(θ) + (1-cos(θ))*z^2  |

rotate(data::Array, vec::PVector, θ::Number; vel::Bool = true)

Rotate all particles in data around direction vector vec by angle θ.

vel::Bool = false: Rotate velocity around origin at the same time.

rotate(p::PVector, vec::PVector, θ::Number, center::PVector)

Rotate p around direction vector vec at center point by angle θ.

rotate(data, vec::PVector, θ::Number, center::PVector; vel::Bool = true)

Rotate all particles in data around direction vector vec at center point by angle θ.

vel::Bool = false: Rotate velocity around origin at the same time.

PhysicalParticles.cylinderial — Function

cylinderial(p::PVector2D)

Convert PVector2D to Cylinderial Coordinates. Returns Tuple(r, θ)

cylinderial(p::PVector)

Convert PVector to Cylinderial Coordinates. Returns Tuple(r, θ, z)

PhysicalParticles.cylinderial2xyz — Function

cylinderial2xyz(r::Number, θ::Number)

Convert Cylinderial Coordinates to PVector2D

cylinderial2xyz(r::Number, θ::Number, z::Number)

Convert Cylinderial Coordinates to PVector

PhysicalParticles.spherical — Function

spherical(p::PVector)

Convert PVector to Spherical Coordinates. Returns Tuple(r, θ, ϕ)

PhysicalParticles.spherical2xyz — Function

spherical2xyz(r::Number, θ::Number, ϕ::Number)

Convert Spherical Coordinates to PVector

Internal

PhysicalParticles.min_x — Function

min_x(a::AbstractPoint, b::AbstractPoint)
min_x(a::AbstractParticle, b::AbstractParticle)

Return the smaller :x field of a and b

PhysicalParticles.max_x — Function

max_x(a::AbstractPoint, b::AbstractPoint)
max_x(a::AbstractParticle, b::AbstractParticle)

Return the larger :x field of a and b

PhysicalParticles.min_y — Function

min_y(a::AbstractPoint, b::AbstractPoint)
min_y(a::AbstractParticle, b::AbstractParticle)

Return the smaller :y field of a and b

PhysicalParticles.max_y — Function

max_y(a::AbstractPoint, b::AbstractPoint)
max_y(a::AbstractParticle, b::AbstractParticle)

Return the larger :y field of a and b

PhysicalParticles.min_z — Function

min_z(a::AbstractPoint, b::AbstractPoint)
min_z(a::AbstractParticle, b::AbstractParticle)

Return the smaller :z field of a and b

PhysicalParticles.max_z — Function

max_z(a::AbstractPoint, b::AbstractPoint)
max_z(a::AbstractParticle, b::AbstractParticle)

Return the larger :z field of a and b

PhysicalParticles.minimum_x — Function

function minimum_x(a::Array{T,N}) where T <: AbstractPoint where N
function minimum_x(a::Array{T,N}) where T <: AbstractParticle where N
function minimum_x(a::StructArray)

Return the minimum :x field of points in array a Return the minimum Pos.x of particles in array a

PhysicalParticles.maximum_x — Function

function maximum_x(a::Array{T,N}) where T <: AbstractPoint where N
function maximum_x(a::Array{T,N}) where T <: AbstractParticle where N
function maximum_x(a::StructArray)

Return the maximum :x field of points in array a Return the maximum Pos.x of particles in array a

PhysicalParticles.minimum_y — Function

function minimum_y(a::Array{T,N}) where T <: AbstractPoint where N
function minimum_y(a::Array{T,N}) where T <: AbstractParticle where N
function minimum_y(a::StructArray)

Return the minimum :y field of points in array a Return the minimum Pos.y of particles in array a

PhysicalParticles.maximum_y — Function

function maximum_y(a::Array{T,N}) where T <: AbstractPoint where N
function maximum_y(a::Array{T,N}) where T <: AbstractParticle where N
function maximum_y(a::StructArray)

Return the maximum :y field of points in array a Return the maximum Pos.y of particles in array a

PhysicalParticles.minimum_z — Function

function minimum_z(a::Array{T,N}) where T <: AbstractPoint where N
function minimum_z(a::Array{T,N}) where T <: AbstractParticle where N
function minimum_z(a::StructArray)

Return the minimum :z field of points in array a Return the minimum Pos.z of particles in array a

PhysicalParticles.maximum_z — Function

function maximum_z(a::Array{T,N}) where T <: AbstractPoint where N
function maximum_z(a::Array{T,N}) where T <: AbstractParticle where N
function maximum_z(a::StructArray)

Return the maximum :z field of points in array a Return the maximum Pos.z of particles in array a